pyproject.toml
[project]
name = "fairly"
version = "0.1.0-beta.3"
authors = [
  { name="Serkan Girgin", email="s.girgin@utwente.nl" },
  { name="Manuel Garcia Alvarez", email="m.g.garciaalvarez@tudelft.nl" },
  { name="Jose Urra Llanusa", email="j.c.urrallanusa@tudelft.nl" }, ]
description = "A package to create, publish, and download research datasets"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent", ]
dependencies = [
    "PyYAML",
    "requests_toolbelt",
    "python-dateutil", 
    "python-dotenv==0.20.0",
    "typer>=0.6.1"
    ]
keywords = ["fairly", "open science", "research data", "data management"]

[project.optional-dependencies]
dev = [
    "pytest",
    "ipython", 
    "pytest-recording==0.12.1",
    "pytest-cov==4.0.0",
    "vcrpy==4.2.1"
    ]

[project.urls]
"Homepage" = "https://github.com/ITC-CRIB/fairly"
"Bug Tracker" = "https://github.com/ITC-CRIB/fairly/issues"
"Documentation" = "https://jupyterfair.readthedocs.io/en/latest/"
"Funding" = "https://nwo.nl/en/researchprogrammes/open-science/open-science-fund"
